Moved WildcardMatcher to SAF.Common to resolve namespace conflict.
