+++ src/Messaging/SAF.Messaging.Routing/SAF.Messaging.Routing.csproj
@@ -12,12 +12,8 @@ SPDX-License-Identifier: MPL-2.0
   </PropertyGroup>
 
   <PropertyGroup>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <AllowUnsafeBlocks>False</AllowUnsafeBlocks>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\..\Communication\SAF.Communication.PubSub\WildcardMatcher.cs" Link="WildcardMatcher.cs" />
-  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
